backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
boto3==1.29.7
email-validator==2.1.0
pyotp==2.9.0
qrcode[pil]==7.4.2
reportlab==4.0.7
httpx==0.25.2
apscheduler==3.10.4

